add user role

+++ routes/upcomingEvent.js
@@ -10,28 +10,34 @@ router.use(bodyParser.json());
 
 const upcomingEvent = require('../model/UpcomingEvent');
 
-router.post('/', (req, res) => verifyToken(req.headers['authorization'], (err) => {
+/** Create an event */
+router.post('/', (req, res) => verifyToken(req.headers['authorization'], (err, decoded) => {
     if(err) {
         if(err === 'invalid token')
             return res.status(403).send('Unauthorized Access');
         return res.status(500).send('Internal Server Error: Unable to verify token');
     }
-    return upcomingEvent.create({
-        date: req.body.date,
-        event: req.body.event
-    }, (err, upcomingEvent) => {
-        if(err)
-            return res.status(500).send('Internal Server Error: Unable to create upcoming events');
-        return res.status(201).send({message: 'Event created!', id: upcomingEvent._id});    
-    });
+    if(decoded.role === 'admin') {
+        return upcomingEvent.create({
+            date: req.body.date,
+            event: req.body.event
+        }, (err, upcomingEvent) => {
+            if(err)
+                return res.status(500).send('Internal Server Error: Unable to create upcoming events');
+            return res.status(201).send({message: 'Event created!', id: upcomingEvent._id});    
+        });
+    }
+    return res.status(403).send('Permission is restricted');
 }));
 
+/** Get events */
 router.get('/', (req, res) => verifyToken(req.headers['authorization'], (err) => {
     if(err) {
         if(err === 'invalid token')
             return res.status(403).send('Unauthorized Access');
         return res.status(500).send('Internal Server Error: Unable to verify token');    
     }
+    //If no query params, return all the events
     if(Object.keys(req.query).length === 0) {
         return upcomingEvent.find({}, (err, events) => {
             if(err)
@@ -39,6 +45,7 @@ router.get('/', (req, res) => verifyToken(req.headers['authorization'], (err) =>
             return res.status(200).send(events);    
         });
     }
+    //Return the event for the given date 
     if(req.query.date) {
         return upcomingEvent.findOne({date: req.query.date}, (err, event) => {
             if(err)
@@ -48,6 +55,7 @@ router.get('/', (req, res) => verifyToken(req.headers['authorization'], (err) =>
     }
 }));
 
+/* Get an event by its ID */
 router.get('/:id', (req, res) => verifyToken(req.headers['authorization'], (err) => {
     if(err) {
         if(err === 'invalid token')
@@ -61,17 +69,21 @@ router.get('/:id', (req, res) => verifyToken(req.headers['authorization'], (err)
     }) ;
 }))
 
-router.delete('/:id', (req, res) => verifyToken(req.headers['authorization'], (err) => {
+/* Delete an event by its ID */
+router.delete('/:id', (req, res) => verifyToken(req.headers['authorization'], (err, decoded) => {
     if(err) {
         if(err === 'invalid token')
             return res.status(403).send('Unauthorized Access');
         return res.status(500).send('Internal Server Error: Unable to verify token');    
     }   
-    return upcomingEvent.findOneAndDelete({_id:req.params.id}, (err, event) => {
-        if(err)
-            return res.status(500).send('Internal Server Error: Unable to find any event');    
-        return res.status(200).send(event);
-    }) ;
+    if(decoded.role === 'admin') {
+        return upcomingEvent.findOneAndDelete({_id:req.params.id}, (err, event) => {
+            if(err)
+                return res.status(500).send('Internal Server Error: Unable to find any event');    
+            return res.status(200).send({id: event._id, status: 'deleted', message: 'record is deleted'});
+        });
+    } 
+    return res.status(403).send('Permission is restricted');
 }))
 
 module.exports = router;
